Refactor imports in training and evaluation scripts
